Merged PR 13470: More image fixes
additional image fixes, we shouldn't use widenImageLoader if the image is not coming from widen, eg. fallback images.

**Build will fail as long as image for "coating -> automatic nozzles" is missing on /categories/coating.
Niki is gonna fix this.** => No longer valid, as we no longer have the hardcoded widenImageLoader for these images.
So builds will no longer fail when an image is missing here. I'll leave it like this for now...